+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 ### Added
+- README file
 - Driver abstract class
 - Driver dedicated for JSON format
 - Changelog parser
@@ -7,4 +8,5 @@
 - Example folder with test script
 - README file
 - Library manager
-- Last version getter
+- Last version getter for JSON Driver
+- All versions getter for JSON Driver